Update mdpvt.p8 file to enhance visual output
Severe changes were made on the mdpvt.p8 file. For starters, sections of underutilized code were removed to streamline the application and enhance its overall performance. Additionally, the graphics data has been improved through the modification of existing hexadecimal patterns. The main aim of this is to depict a more accurate visual output in the application, which was made possible through the adjustment of colour values and matrices.