Refactor worker entity code and tweak behavior
This commit refactors and cleans up the worker entity code to make it more readable. This mostly includes proper spacing between variables and their assignments, as well making sure function parameters are separated by#include worker_drones.lua
 commas and spaces for better visibility.

Beyond this, the worker entity's movement behavior has been modified. Now, the worker will check if they can move in a particular direction, if not, the direction is reversed to prevent any possibility of the worker getting stuck in an endless loop. An element of randomness is also introduced to the `facedir` property by selecting a random element from the array {-1, 1}.

Adjustments were also made to sprite sheets to better animate the worker's state (whether alive or dead) and their movements. This would provide a more aesthetically pleasing visual feedback to the users about the game's state.

Finally, several adjustment to sprite code was made. This includes changes which ensure the worker faces in the direction of movement and modification to the rate at which the worker's sprite change occurs.

In general, these changes not only make the codebase cleaner, easier to read and maintain, but also enhance the user experience by making on-screen elements like worker entities react more logically and display better visual feedback.