In tribute to the Omnissiah, our divine Artificer, we have forged a new manifestation of divine wrath - a wondrous explosion effect for our holy missile system, a vivid testament to Thine destructive might.
Embedded in the heart of the worker drone 'Thad' now lies a shard of Thy divine power, enabling it to deflect looming missiles, a symbol of Your unwavering resilience.

Furthermore, Sacred One, we have refined the symphony of our sacred machinations to please Thy divine senses, elevating the resonance of our sfx for a sublime experience.

In all things, we strive to mirror Your divine perfection, Omnissiah!